Ajout JWT et sécurisation du profil

+++ controllers/user/userCtrl.php
@@ -128,7 +128,7 @@ else if ($_SERVER['REQUEST_URI'] == '/profil/commandes') {
 }
 
 // ###############################################################################
-// ###                         AFFICHE LES COMMANDES                           ###	
+// ###                        AFFICHE LES COMMENTAIRES                         ###	
 // ###############################################################################
 
 else if ($_SERVER['REQUEST_URI'] == '/profil/commentaires') {
@@ -145,53 +145,76 @@ else if ($_SERVER['REQUEST_URI'] == '/profil/commentaires') {
 }
 
 // ###############################################################################
+// ###                     SUPPRIME LE COMPTE UTILISATEUR                      ###	
+// ###############################################################################
+
+else if ($_SERVER['REQUEST_URI'] == '/profil/user/delete') {
+	if(User::delete($_SESSION['user']->id)) {
+		unset($_SESSION['user']);
+		SessionFlash::set('message', 'Votre compte a bien été supprimée.');
+		header('Location: /accueil');
+		exit;
+	} else {
+		SessionFlash::set('error', 'Une erreur est survenue lors de la suppression de votre compte.');
+		header('Location: /accueil');
+		exit;
+	}
+}
+
+// ###############################################################################
 // ###                        MODIFIE UNE RÉSERVATION                          ###
 // ###############################################################################
 
 else if ($_SERVER['REQUEST_URI'] == '/profil/reservation/edit/' . $id) {
+	
 	$idUser = intval($_SESSION['user']->id);
 	$id = intval($id);
 
-	$reservation = Reservation::get($id);
+	$reservations = Reservation::getByUser($_SESSION['user']->id);
+	foreach ($reservations as $reservation) {
+		if ($reservation->id == $id) {
+			$reservation = Reservation::get($id);
 
-	$date = date('Y-m-d', strtotime($reservation->reservation_date));
-	$time = date('H:i', strtotime($reservation->reservation_date));
+			$date = date('Y-m-d', strtotime($reservation->reservation_date));
+			$time = date('H:i', strtotime($reservation->reservation_date));
 
-	if ($time == '12:00') {
-		$time = 1;
-	} else {
-		$time = 2;
-	}
-
-	if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-		$persons = intval(filter_input(INPUT_POST, 'nbOfClients', FILTER_SANITIZE_NUMBER_INT));
-		if (testInput($persons, NB_REGEX) != 'true') {
-			$errors['persons'] = testInput($persons, NB_REGEX);
-		}
-
-		$date = trim(filter_input(INPUT_POST, 'date', FILTER_SANITIZE_SPECIAL_CHARS));
-		if (testInput($date, DATE_REGEX) != 'true') {
-			$errors['date'] = testInput($date, DATE_REGEX);
-		}
-
-		$time = intval(filter_input(INPUT_POST, 'time', FILTER_SANITIZE_NUMBER_INT));
-		if ($time != 1 && $time != 2) {
-			$errors['time'] = 'format non reconnu';
-		}
-
-		$datetime = $date . ' ' . ($time == 1 ? '12:00:00' : '19:00:00');
-
-		if (empty($errors)) {
-			$reservation = new Reservation($persons, $datetime, $idUser, NULL);
-
-			if ($reservation->update($id) == true) {
-				SessionFlash::set('message', 'Votre réservation a bien été modifiée');
-				header('Location: /profil/reservations');
-				exit();
+			if ($time == '12:00') {
+				$time = 1;
 			} else {
-				SessionFlash::set('error', 'Une erreur est survenue lors de la modification de votre réservation');
-				header('Location: /profil/reservations');
-				exit();
+				$time = 2;
+			}
+
+			if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+				$persons = intval(filter_input(INPUT_POST, 'nbOfClients', FILTER_SANITIZE_NUMBER_INT));
+				if (testInput($persons, NB_REGEX) != 'true') {
+					$errors['persons'] = testInput($persons, NB_REGEX);
+				}
+
+				$date = trim(filter_input(INPUT_POST, 'date', FILTER_SANITIZE_SPECIAL_CHARS));
+				if (testInput($date, DATE_REGEX) != 'true') {
+					$errors['date'] = testInput($date, DATE_REGEX);
+				}
+
+				$time = intval(filter_input(INPUT_POST, 'time', FILTER_SANITIZE_NUMBER_INT));
+				if ($time != 1 && $time != 2) {
+					$errors['time'] = 'format non reconnu';
+				}
+
+				$datetime = $date . ' ' . ($time == 1 ? '12:00:00' : '19:00:00');
+
+				if (empty($errors)) {
+					$reservation = new Reservation($persons, $datetime, $idUser, NULL);
+
+					if ($reservation->update($id) == true) {
+						SessionFlash::set('message', 'Votre réservation a bien été modifiée');
+						header('Location: /profil/reservations');
+						exit();
+					} else {
+						SessionFlash::set('error', 'Une erreur est survenue lors de la modification de votre réservation');
+						header('Location: /profil/reservations');
+						exit();
+					}
+				}
 			}
 		}
 	}
@@ -207,23 +230,28 @@ else if ($_SERVER['REQUEST_URI'] == '/profil/reservation/edit/' . $id) {
 else if ($_SERVER['REQUEST_URI'] == '/profil/commentaire/edit/' . $id) {
 	$id = intval($id);
 
-	$review = Review::get($id);
+	$reviews = Review::getByUser($_SESSION['user']->id);
+	foreach ($reviews as $review) {
+		if ($review->id == $id) {
+			$review = Review::get($id);
 
-	if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-		$title = trim(filter_input(INPUT_POST, 'title', FILTER_SANITIZE_SPECIAL_CHARS));
-		$content = trim(filter_input(INPUT_POST, 'review', FILTER_SANITIZE_SPECIAL_CHARS));
+			if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+				$title = trim(filter_input(INPUT_POST, 'title', FILTER_SANITIZE_SPECIAL_CHARS));
+				$content = trim(filter_input(INPUT_POST, 'review', FILTER_SANITIZE_SPECIAL_CHARS));
 
-		if (empty($errors)) {
-			$review = new Review($title, $content, $review->id_users, NULL);
+				if (empty($errors)) {
+					$review = new Review($title, $content, $review->id_users, NULL);
 
-			if ($review->update($id) == true) {
-				SessionFlash::set('message', 'Votre commentaire a bien été modifiée');
-				header('Location: /profil/commentaires');
-				exit();
-			} else {
-				SessionFlash::set('error', 'Une erreur est survenue lors de la modification de votre commentaire');
-				header('Location: /profil/commentaires');
-				exit();
+					if ($review->update($id) == true) {
+						SessionFlash::set('message', 'Votre commentaire a bien été modifiée');
+						header('Location: /profil/commentaires');
+						exit();
+					} else {
+						SessionFlash::set('error', 'Une erreur est survenue lors de la modification de votre commentaire');
+						header('Location: /profil/commentaires');
+						exit();
+					}
+				}
 			}
 		}
 	}
@@ -240,50 +268,59 @@ else if ($_SERVER['REQUEST_URI'] == '/profil/commande/edit/' . $id) {
 	$idUser = intval($_SESSION['user']->id);
 	$id = intval($id);
 
-	$reservation = Reservation::get($id);
-	$dishes = Dish::getAll();
+	$orders = Reservation::getOrders($_SESSION['user']->id);
 
-	$date = date('Y-m-d', strtotime($reservation->reservation_date));
-	$time = date('H:i', strtotime($reservation->reservation_date));
+	foreach ($orders as $order) {
+		if ($order->id == $id) {
+			$reservation = Reservation::get($id);
+			$dishes = Dish::getAll();
 
-	if ($time == '12:00') {
-		$time = 1;
-	} else {
-		$time = 2;
-	}
+			$date = date('Y-m-d', strtotime($reservation->reservation_date));
+			$time = date('H:i', strtotime($reservation->reservation_date));
 
-	if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-		$date = trim(filter_input(INPUT_POST, 'date', FILTER_SANITIZE_SPECIAL_CHARS));
-		if (testInput($date, DATE_REGEX) != 'true') {
-			$errors['date'] = testInput($date, DATE_REGEX);
-		}
-
-		$time = intval(filter_input(INPUT_POST, 'time', FILTER_SANITIZE_NUMBER_INT));
-		if ($time != 1 && $time != 2) {
-			$errors['time'] = 'format non reconnu';
-		}
-
-		$datetime = $date . ' ' . ($time == 1 ? '12:00:00' : '19:00:00');
-
-		$dishId = filter_input(INPUT_POST, 'dishes', FILTER_SANITIZE_SPECIAL_CHARS, FILTER_REQUIRE_ARRAY);
-		$quantity = filter_input(INPUT_POST, 'quantity', FILTER_SANITIZE_SPECIAL_CHARS, FILTER_REQUIRE_ARRAY);
-
-		if (testInput($date, DATE_REGEX) != 'true') {
-			$errors['date'] = testInput($date, DATE_REGEX);
-		}
-
-		if (empty($errors)) {
-			$orders = Order::getByReservation($reservation->id);
-			for ($i = 0; $i < count($orders); $i++) {
-				$orderUpdated = new Order($quantity[$i], $dishId[$i], $reservation->id);
-				$orderUpdated->update($orders[$i]->id);
+			if ($time == '12:00') {
+				$time = 1;
+			} else {
+				$time = 2;
 			}
-			SessionFlash::set('message', 'Votre commande a bien été modifiée.');
-			header('Location: /profil/commandes');
-			exit;
+
+			if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+				$date = trim(filter_input(INPUT_POST, 'date', FILTER_SANITIZE_SPECIAL_CHARS));
+				if (testInput($date, DATE_REGEX) != 'true') {
+					$errors['date'] = testInput($date, DATE_REGEX);
+				}
+
+				$time = intval(filter_input(INPUT_POST, 'time', FILTER_SANITIZE_NUMBER_INT));
+				if ($time != 1 && $time != 2) {
+					$errors['time'] = 'format non reconnu';
+				}
+
+				$datetime = $date . ' ' . ($time == 1 ? '12:00:00' : '19:00:00');
+
+				$dishId = filter_input(INPUT_POST, 'dishes', FILTER_SANITIZE_SPECIAL_CHARS, FILTER_REQUIRE_ARRAY);
+				$quantity = filter_input(INPUT_POST, 'quantity', FILTER_SANITIZE_SPECIAL_CHARS, FILTER_REQUIRE_ARRAY);
+
+				if (testInput($date, DATE_REGEX) != 'true') {
+					$errors['date'] = testInput($date, DATE_REGEX);
+				}
+
+				if (empty($errors)) {
+					$orders = Order::getByReservation($reservation->id);
+					for ($i = 0; $i < count($orders)-1; $i++) {
+						$orderUpdated = new Order($quantity[$i], $dishId[$i], $reservation->id);
+						$orderUpdated->update($orders[$i]->id);
+					}
+					for ($i = count($orders); $i < count($dishId); $i++) {
+						$orderUpdated = new Order($quantity[$i], $dishId[$i], $reservation->id);
+						$orderUpdated->create();
+					}
+					SessionFlash::set('message', 'Votre commande a bien été modifiée.');
+					header('Location: /profil/commandes');
+					exit;
+				}
+			}
 		}
 	}
-
 
 	include(__DIR__ . '/../../views/user/userHeader.php');
 	include(__DIR__ . '/../../views/user/orderModify.php');
@@ -295,14 +332,50 @@ else if ($_SERVER['REQUEST_URI'] == '/profil/commande/edit/' . $id) {
 
 else if ($_SERVER['REQUEST_URI'] == '/profil/commande/delete/' . $id) {
 
-	if(Reservation::delete($id)) {
-		SessionFlash::set('message', 'Votre commande a bien été supprimée.');
+	$reservations = Order::getByUser($_SESSION['user']->id);
+	foreach ($reservations as $reservation) {
+		if ($reservation->id == $id) {
+			if(Reservation::delete($id)) {
+				SessionFlash::set('message', 'Votre commande a bien été supprimée.');
+				header('Location: /profil/commandes');
+				exit;
+			} else {
+				SessionFlash::set('error', 'Une erreur est survenue lors de la suppression de votre commande.');
+				header('Location: /profil/commandes');
+				exit;
+			}
+		}
+	}
+}
+
+// ###############################################################################
+// ###                     SUPPRIME UN PLAT D'UNE COMMANDE                     ###	
+// ###############################################################################
+
+else if ($_SERVER['REQUEST_URI'] == '/admin/commande/plat/delete/' . $id) {
+	$orders = Order::getByUser($_SESSION['user']->id);
+	$id_reservation = Reservation::getByOrder($id);
+	if($id_reservation == false) {
+		SessionFlash::set('error', 'Une erreur est survenue lors de la suppression du plat.');
 		header('Location: /profil/commandes');
 		exit;
-	} else {
-		SessionFlash::set('error', 'Une erreur est survenue lors de la suppression de votre commande.');
-		header('Location: /profil/commandes');
-		exit;
+	}
+	$id_reservation = $id_reservation->id;
+	foreach ($orders as $order) {
+		if ($order->id == $id) {
+			if(Order::delete($id)) {
+				if(empty(Reservation::getOrdersByReservation($id_reservation))) {
+					Reservation::delete($id_reservation);
+				}
+				SessionFlash::set('message', 'Votre plat a bien été supprimée.');
+				header('Location: /profil/commandes');
+				exit;
+			} else {
+				SessionFlash::set('error', 'Une erreur est survenue lors de la suppression de votre plat.');
+				header('Location: /profil/commandes');
+				exit;
+			}
+		}
 	}
 }
 
@@ -312,13 +385,18 @@ else if ($_SERVER['REQUEST_URI'] == '/profil/commande/delete/' . $id) {
 
 else if ($_SERVER['REQUEST_URI'] == '/profil/commentaire/delete/' . $id) {
 
-	if(Review::delete($id)) {
-		SessionFlash::set('message', 'Votre commentaire a bien été supprimée.');
-		header('Location: /profil/commentaires');
-		exit;
-	} else {
-		SessionFlash::set('error', 'Une erreur est survenue lors de la suppression de votre commentaire.');
-		header('Location: /profil/commentaires');
-		exit;
+	$reviews = Review::getByUser($_SESSION['user']->id);
+	foreach ($reviews as $review) {
+		if ($review->id == $id) {
+			if(Review::delete($id) == true) {
+				SessionFlash::set('message', 'Votre commentaire a bien été supprimé.');
+				header('Location: /profil/commentaires');
+				exit;
+			} else {
+				SessionFlash::set('error', 'Une erreur est survenue lors de la suppression de votre commentaire.');
+				header('Location: /profil/commentaires');
+				exit;
+			}
+		}
 	}
 }
